Rename startpage folder
